1. Added route setup for the maintenance page in App.jsx. 2. Added a dashboard menu button for the maintenance page in DashboardMenu. 3. Moved content with CSS from the Incidents page to the Maintenance page. 4. Updated pathMap in PathMap.js to include the maintenance route.
